+++ nel/tools/3d/zone_welder/zone_welder.cpp
@@ -26,15 +26,15 @@
 
 #include "../zone_lib/zone_utility.h"
 
-#include <iostream.h>
+#include <iostream>
 #include "nel/misc/types_nl.h"
 #include "nel/misc/file.h"
 #include "nel/3d/quad_tree.h"
-#include "nel/../../src/3d/zone.h"
-#include "nel/../../src/3d/landscape.h"
-#include "nel/../../src/3d/zone_smoother.h"
-#include "nel/../../src/3d/zone_tgt_smoother.h"
-#include "nel/../../src/3d/zone_corner_smoother.h"
+#include "nel/3d/zone.h"
+#include "nel/3d/landscape.h"
+#include "nel/3d/zone_smoother.h"
+#include "nel/3d/zone_tgt_smoother.h"
+#include "nel/3d/zone_corner_smoother.h"
 #include <vector>
 #include <set>
 
@@ -678,7 +678,7 @@ void weldZones(const char *center)
 				uint	idVert= pa.BaseVertices[corner];
 
 				// write this vertex in array.
-				centerVertices.resize( max(centerVertices.size(), idVert+1) );
+				centerVertices.resize( max((uint)centerVertices.size(), idVert+1) );
 				centerVertices[idVert]= pa.Patch.Vertices[corner];
 			}
 		}
@@ -703,7 +703,7 @@ void weldZones(const char *center)
 					uint	idVert= pa.BaseVertices[corner];
 
 					// write this vertex in array.
-					verts.resize( max(verts.size(), idVert+1) );
+					verts.resize( max((uint)verts.size(), idVert+1) );
 					verts[idVert].Vertex= pa.Patch.Vertices[corner];
 					verts[idVert].OnBorder= false;
 				}
@@ -1005,5 +1005,6 @@ int main(sint argc, char **argv)
 
 	return 0;
 }
+
